requirements.txt
requests
PyCryptodome